Deprecate lsr and rmr command to be alias for "ls -R" and "rm -R".
