+++ shell/src/main/java/tachyon/shell/command/LsCommand.java
@@ -16,18 +16,29 @@
 package tachyon.shell.command;
 
 import java.io.IOException;
+import java.util.Collections;
+import java.util.Comparator;
+import java.util.List;
 
 import javax.annotation.concurrent.ThreadSafe;
 
+import org.apache.commons.cli.CommandLine;
+import org.apache.commons.cli.Option;
+import org.apache.commons.cli.Options;
+
+import tachyon.Constants;
 import tachyon.TachyonURI;
 import tachyon.client.file.FileSystem;
+import tachyon.client.file.URIStatus;
 import tachyon.conf.TachyonConf;
+import tachyon.exception.TachyonException;
+import tachyon.util.FormatUtils;
 
 /**
  * Displays information for all directories and files directly under the path specified in args.
  */
 @ThreadSafe
-public final class LsCommand extends AbstractLsCommand {
+public final class LsCommand extends WithWildCardPathCommand {
 
   /**
    * Constructs a new instance to display information for all directories and files directly under
@@ -46,13 +57,86 @@ public final class LsCommand extends AbstractLsCommand {
   }
 
   @Override
-  void runCommand(TachyonURI path) throws IOException {
-    ls(path, false);
+  protected int getNumOfArgs() {
+    return 1;
+  }
+
+  @Override
+  protected Options getOptions() {
+    Options opts = new Options();
+    // Add R option for recursively.
+    Option recursive = Option.builder("R")
+        .required(false)
+        .hasArg(false)
+        .desc("recusively")
+        .build();
+
+    opts.addOption(recursive);
+    return opts;
+  }
+
+  /**
+   * Displays information for all directories and files directly under the path specified in args.
+   *
+   * @param path The {@link TachyonURI} path as the input of the command
+   * @param recursive Whether list the path recursively
+   * @throws IOException if a non-Tachyon related exception occurs
+   */
+  private void ls(TachyonURI path, boolean recursive) throws IOException {
+    List<URIStatus> statuses = listStatusSortedByIncreasingCreationTime(path);
+    for (URIStatus status : statuses) {
+      String inMemory = "";
+      if (!status.isFolder()) {
+        if (100 == status.getInMemoryPercentage()) {
+          inMemory = "In Memory";
+        } else {
+          inMemory = "Not In Memory";
+        }
+      }
+      System.out.format(Constants.COMMAND_FORMAT_LS,
+          FormatUtils.formatPermission((short) status.getPermission(), status.isFolder()),
+          status.getUserName(), status.getGroupName(),
+          FormatUtils.getSizeFromBytes(status.getLength()),
+          CommandUtils.convertMsToDate(status.getCreationTimeMs()), inMemory, status.getPath());
+      if (recursive && status.isFolder()) {
+        ls(new TachyonURI(path.getScheme(), path.getAuthority(), status.getPath()), true);
+      }
+    }
+  }
+
+  private List<URIStatus> listStatusSortedByIncreasingCreationTime(TachyonURI path)
+      throws IOException {
+    List<URIStatus> statuses;
+    try {
+      statuses = mFileSystem.listStatus(path);
+    } catch (TachyonException e) {
+      throw new IOException(e.getMessage());
+    }
+    Collections.sort(statuses, new Comparator<URIStatus>() {
+      @Override
+      public int compare(URIStatus status1, URIStatus status2) {
+        long t1 = status1.getCreationTimeMs();
+        long t2 = status2.getCreationTimeMs();
+        if (t1 < t2) {
+          return -1;
+        }
+        if (t1 == t2) {
+          return 0;
+        }
+        return 1;
+      }
+    });
+    return statuses;
+  }
+
+  @Override
+  public void runCommand(TachyonURI path, CommandLine cl) throws IOException {
+    ls(path, cl.hasOption("R"));
   }
 
   @Override
   public String getUsage() {
-    return "ls <path>";
+    return "ls [-R] <path>";
   }
 
   @Override
