+++ shell/src/main/java/tachyon/shell/command/ReportCommand.java
@@ -19,6 +19,8 @@ import java.io.IOException;
 
 import javax.annotation.concurrent.ThreadSafe;
 
+import org.apache.commons.cli.CommandLine;
+
 import tachyon.TachyonURI;
 import tachyon.client.file.FileSystem;
 import tachyon.client.lineage.LineageFileSystem;
@@ -45,7 +47,7 @@ public final class ReportCommand extends WithWildCardPathCommand {
   }
 
   @Override
-  void runCommand(TachyonURI path) throws IOException {
+  void runCommand(TachyonURI path, CommandLine cl) throws IOException {
     try {
       LineageFileSystem.get().reportLostFile(path);
       System.out.println(path + " has reported been reported as lost.");
